requirements.txt
asgiref==3.11.1
attrs==26.1.0
colorama==0.4.6
coverage==7.13.5
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.14
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mutagen==1.47.0
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
